Refactor: Rename AuthenticationUseCase package
