+++ presentation/src/main/java/com/london/presentation/feature/account/AccountViewModel.kt
@@ -6,7 +6,7 @@ import com.london.domain.contentrestriction.ContentRestrictionLevel
 import com.london.domain.language.AppLanguage
 import com.london.domain.theme.AppTheme
 import com.london.domain.usecase.GetAccountDetails
-import com.london.domain.usecase.login.AuthenticationUseCase
+import com.london.domain.usecase.authentication.AuthenticationUseCase
 import com.london.presentation.feature.account.state.AccountUiState
 import com.london.presentation.shared.base.BaseViewModel
 import com.london.presentation.shared.base.ErrorState
